+++ apps/judicial-system/backend/src/app/pdf/formatters.ts
